# apps/api/src/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IdeiaERP Commerce Sync API",
    "version": "1.0.0",
    "description": "API REST para sincronização entre IdeiaERP e lojas virtuais"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "UUID",
        "description": "Token de autenticação obtido no endpoint /auth/login"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "usuario": {
            "type": "object",
            "properties": {
              "usuario_id": {
                "type": "number"
              },
              "nome": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "token": {
            "type": "string",
            "description": "Token de acesso (UUID)"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token para renovar o token de acesso (UUID)"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de expiração do token"
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout de usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso"
          },
          "401": {
            "description": "Não autenticado"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Obter informações do usuário logado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Informações do usuário"
          },
          "401": {
            "description": "Não autenticado"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Renovar token de acesso",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "Refresh token para renovar o token de acesso"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token renovado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido ou expirado"
          }
        }
      }
    },
...
  }
}